CORE-237 moved multiple autocomplete module to libs
