+++ frontend/apps/quarano-frontend/src/app/modules/basic-data/basic-data.module.ts
@@ -6,7 +6,7 @@ import { CommonModule } from '@angular/common';
 import { BasicDataComponent } from './basic-data.component';
 import { BasicDataRoutingModule } from './basic-data-routing.module';
 import { InitialQuestionaireFormComponent } from './initial-questionaire-form/initial-questionaire-form.component';
-import { MultipleAutocompleteModule } from '../../ui/multiple-autocomplete/multiple-autocomplete.module';
+import { SharedUiMultipleAutocompleteModule } from '@qro/shared/ui-multiple-autocomplete';
 import { SharedUiConfirmationDialogModule } from '@qro/shared/ui-confirmation-dialog';
 import { MyClientDataResolver } from '../../resolvers/my-client-data.resolver';
 import { EncountersResolver } from '../../resolvers/encounters.resolver';
@@ -21,7 +21,7 @@ import { SharedUiMaterialModule } from '@qro/shared/ui-material';
     SharedUiMaterialModule,
     FormsModule,
     ReactiveFormsModule,
-    MultipleAutocompleteModule,
+    SharedUiMultipleAutocompleteModule,
     AppFormsModule,
     SharedUiConfirmationDialogModule,
     SharedUiButtonModule
