+++ frontend/libs/shared/ui-multiple-autocomplete/README.md
@@ -1,0 +1,7 @@
+# shared-ui-multiple-autocomplete
+
+This library was generated with [Nx](https://nx.dev).
+
+## Running unit tests
+
+Run `nx test shared-ui-multiple-autocomplete` to execute the unit tests.
