+++ frontend/libs/shared/ui-multiple-autocomplete/src/index.ts
@@ -1,0 +1,1 @@
+export * from './lib/shared-ui-multiple-autocomplete.module';
