+++ frontend/libs/shared/ui-multiple-autocomplete/src/lib/multiple-autocomplete/multiple-autocomplete.component.scss
