+++ frontend/libs/shared/ui-multiple-autocomplete/src/lib/shared-ui-multiple-autocomplete.module.spec.ts
@@ -1,0 +1,14 @@
+import { async, TestBed } from '@angular/core/testing';
+import { SharedUiMultipleAutocompleteModule } from './shared-ui-multiple-autocomplete.module';
+
+describe('SharedUiMultipleAutocompleteModule', () => {
+  beforeEach(async(() => {
+    TestBed.configureTestingModule({
+      imports: [SharedUiMultipleAutocompleteModule],
+    }).compileComponents();
+  }));
+
+  it('should create', () => {
+    expect(SharedUiMultipleAutocompleteModule).toBeDefined();
+  });
+});
