+++ frontend/libs/shared/ui-multiple-autocomplete/src/lib/shared-ui-multiple-autocomplete.module.ts
@@ -1,8 +1,8 @@
 import { ReactiveFormsModule } from '@angular/forms';
+import { SharedUiMaterialModule } from '@qro/shared/ui-material';
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { MultipleAutocompleteComponent } from './multiple-autocomplete.component';
-import { SharedUiMaterialModule } from '@qro/shared/ui-material';
+import { MultipleAutocompleteComponent } from './multiple-autocomplete/multiple-autocomplete.component';
 
 @NgModule({
   imports: [
@@ -13,4 +13,4 @@ import { SharedUiMaterialModule } from '@qro/shared/ui-material';
   declarations: [MultipleAutocompleteComponent],
   exports: [MultipleAutocompleteComponent]
 })
-export class MultipleAutocompleteModule { }
+export class SharedUiMultipleAutocompleteModule { }
