+++ frontend/libs/shared/ui-multiple-autocomplete/src/test-setup.ts
@@ -1,0 +1,1 @@
+import 'jest-preset-angular';
